config/requirements.txt
asgiref==3.8.1
Django==5.1.5
django-polymorphic==3.1.0
djangorestframework==3.15.2
pillow==11.1.0
python-decouple==3.8
setuptools==75.8.0
sqlparse==0.5.3
